Add files via upload
Here I attached the screenshots of Usecase, UI and Sequence Diagram. However, for the sequence diagram, it was not possible to get the screenshot, because if I zoom out and screenshot, it's not visible enough to read. As a result, the sequence diagram file needs to open by QSEE Superlite software.